Log displacement based on geofences
With the location manager geofencing trick the passive gait task uses to activate passive motion sensor logging, it looks like maybe we can piggyback on that to get better passive displacement data by logging a displacement to Bridge whenever we set a geofence (vs. using significant location changes, which are unpredictable and undependable). We’ll log it to a separate schema so the data analysts can compare the two methods and see which, if either, really is better.